Allow respond option to be false
Prior, if `respond` was falsy, the AnvilConnectExpress constructor
would ignore its value and instead use `true`. This made it impossible
to use `respond` to change the behaviour such that the default error
handler would not be used.

This change makes the check more specific by instead checking whether
or not `respond` is undefined. If it is, `true` is used as a default
just as before. If it is not undefined, then its value is respected.